refactor(core): split progression into sibling modules with facade
Extract xp_levels, progression_hp, asi, class_progression, and
level_up_resolve; keep core.progression as stable re-export facade.
