+++ tests/test_character.py
@@ -151,7 +151,7 @@ def test_starting_max_hp_and_hardcore(
     assert saved.current_hp == saved.max_hp
     stats["constitution"] = 14
     monkeypatch.setattr(
-        "core.progression.roll",
+        "core.progression_hp.roll",
         lambda count, sides, modifier=0: 5 + modifier,
     )
     hard_character = build_new_character(
@@ -175,7 +175,7 @@ def test_hardcore_l1_hp_floor_on_create(
     stats = dict.fromkeys(STAT_NAMES, 10)
     stats["constitution"] = 8  # модификатор −1
     monkeypatch.setattr(
-        "core.progression.roll",
+        "core.progression_hp.roll",
         lambda count, sides, modifier=0: 1 + modifier,
     )
     character = build_new_character(
